refactor: co-locate types with services and add request validation
Split monolithic types.go by moving domain types, constants, and request
structs into *_types.go files alongside their service files. Add
client-side request validation using ozzo-validation for all mutating
operations (Create, Update, Action). Wire Validate() calls into service
methods before HTTP requests.

- instances.go + instances_types.go
- clusters.go + clusters_types.go
- volumes.go + volumes_types.go
- container_deployments.go + container_deployments_types.go
- serverless_jobs.go + serverless_jobs_types.go
- types.go now only contains shared FlexibleFloat helper
- validation.go contains shared IsLatestTag helper
- Move tests to match new file locations

Made-with: Cursor
