+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/verda-cloud/verdacloud-sdk-go
 
 go 1.25
+
+require github.com/go-ozzo/ozzo-validation/v4 v4.3.0 // indirect
