+++ pkg/verda/clusters.go
@@ -29,6 +29,10 @@ func (s *ClusterService) GetByID(ctx context.Context, id string) (*Cluster, erro
 }
 
 func (s *ClusterService) Create(ctx context.Context, req CreateClusterRequest) (*CreateClusterResponse, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+
 	if req.LocationCode == "" {
 		req.LocationCode = LocationFIN01
 	}
