+++ pkg/verda/container_types.go
@@ -5,6 +5,22 @@ import (
 	"net/url"
 )
 
+// ContainerType represents a serverless container compute resource option
+type ContainerType struct {
+	ID                  string         `json:"id"`
+	Model               string         `json:"model"`
+	Name                string         `json:"name"`
+	InstanceType        string         `json:"instance_type"`
+	CPU                 InstanceCPU    `json:"cpu"`
+	GPU                 InstanceGPU    `json:"gpu"`
+	GPUMemory           InstanceMemory `json:"gpu_memory"`
+	Memory              InstanceMemory `json:"memory"`
+	ServerlessPrice     FlexibleFloat  `json:"serverless_price"`
+	ServerlessSpotPrice FlexibleFloat  `json:"serverless_spot_price"`
+	Currency            string         `json:"currency"`
+	Manufacturer        string         `json:"manufacturer"`
+}
+
 type ContainerTypesService struct {
 	client *Client
 }
