+++ pkg/verda/instance_types.go
@@ -6,6 +6,38 @@ import (
 	"net/url"
 )
 
+// InstanceTypeInfo represents detailed instance type information with pricing
+type InstanceTypeInfo struct {
+	ID              string          `json:"id"`
+	InstanceType    string          `json:"instance_type"`
+	Model           string          `json:"model"`
+	Name            string          `json:"name"`
+	CPU             InstanceCPU     `json:"cpu"`
+	GPU             InstanceGPU     `json:"gpu"`
+	GPUMemory       InstanceMemory  `json:"gpu_memory"`
+	Memory          InstanceMemory  `json:"memory"`
+	PricePerHour    FlexibleFloat   `json:"price_per_hour"`
+	SpotPrice       FlexibleFloat   `json:"spot_price"`
+	DynamicPrice    FlexibleFloat   `json:"dynamic_price"`
+	MaxDynamicPrice FlexibleFloat   `json:"max_dynamic_price"`
+	Storage         InstanceStorage `json:"storage"`
+	Currency        string          `json:"currency"`
+	Manufacturer    string          `json:"manufacturer"`
+	BestFor         []string        `json:"best_for"`
+	Description     string          `json:"description"`
+}
+
+// PriceHistoryRecord represents a single price record in the price history
+type PriceHistoryRecord struct {
+	Date                string        `json:"date"`
+	FixedPricePerHour   FlexibleFloat `json:"fixed_price_per_hour"`
+	DynamicPricePerHour FlexibleFloat `json:"dynamic_price_per_hour"`
+	Currency            string        `json:"currency"`
+}
+
+// InstanceTypePriceHistory maps instance type names to their price history records
+type InstanceTypePriceHistory map[string][]PriceHistoryRecord
+
 type InstanceTypesService struct {
 	client *Client
 }
