+++ pkg/verda/instances_test.go
@@ -176,6 +176,7 @@ func TestInstanceService_CreateSpotWithDiscontinuePolicy(t *testing.T) {
 			InstanceType: "CPU.4V.16G",
 			Image:        "ubuntu-24.04",
 			Hostname:     "test-spot-instance",
+			Description:  "test spot instance",
 			SSHKeyIDs:    []string{"key_123"},
 			LocationCode: "FIN-03",
 			IsSpot:       true,
@@ -201,6 +202,7 @@ func TestInstanceService_CreateSpotWithDiscontinuePolicy(t *testing.T) {
 			InstanceType: "CPU.4V.16G",
 			Image:        "ubuntu-24.04",
 			Hostname:     "test-spot-volumes",
+			Description:  "test spot volumes",
 			SSHKeyIDs:    []string{"key_123"},
 			LocationCode: "FIN-03",
 			IsSpot:       true,
