+++ pkg/verda/locations.go
@@ -2,6 +2,18 @@ package verda
 
 import "context"
 
+// Location represents a datacenter location
+type Location struct {
+	Code        string `json:"code"`
+	Name        string `json:"name"`
+	CountryCode string `json:"country_code"`
+}
+
+// Default location (used when no location is specified)
+const (
+	LocationFIN01 = "FIN-01"
+)
+
 type LocationService struct {
 	client *Client
 }
