+++ pkg/verda/validation.go
@@ -1,21 +1,22 @@
+// validation.go contains shared validation helpers used across multiple service
+// type files (e.g. container_deployments_types.go, serverless_jobs_types.go).
+//
+// Service-specific validation logic (Validate methods, validateCreate* functions)
+// lives in the corresponding *_types.go file. Only put helpers here when they
+// are reused by more than one service domain.
+
 package verda
 
 import "strings"
 
-// isLatestTag checks if a container image uses the "latest" tag
-// The API does not allow "latest" tag - a specific version must be used
-func isLatestTag(image string) bool {
-	// Check for explicit :latest tag
+// IsLatestTag checks if a container image uses the "latest" tag.
+// The API does not allow "latest" tag - a specific version must be used.
+func IsLatestTag(image string) bool {
 	if strings.HasSuffix(image, ":latest") {
 		return true
 	}
-
-	// Check for image without any tag (defaults to latest)
-	// e.g., "nginx" without ":tag" would default to "nginx:latest"
 	if !strings.Contains(image, ":") && !strings.Contains(image, "@") {
-		// Image has no tag and no digest - would default to latest
 		return true
 	}
-
 	return false
 }
