+++ pkg/verda/volumes.go
@@ -44,6 +44,9 @@ func (s *VolumeService) GetVolume(ctx context.Context, id string) (*Volume, erro
 }
 
 func (s *VolumeService) CreateVolume(ctx context.Context, req VolumeCreateRequest) (string, error) {
+	if err := req.Validate(); err != nil {
+		return "", err
+	}
 	if req.LocationCode == "" {
 		req.LocationCode = LocationFIN01
 	}
@@ -85,7 +88,9 @@ func (s *VolumeService) DeleteVolume(ctx context.Context, id string, force bool)
 
 // AttachVolume attaches a volume - instance must be shut down first
 func (s *VolumeService) AttachVolume(ctx context.Context, volumeID string, req VolumeAttachRequest) error {
-	// Use map to combine action fields with instance_id
+	if err := req.Validate(); err != nil {
+		return err
+	}
 	payload := map[string]interface{}{
 		"id":          volumeID,
 		"action":      VolumeActionAttach,
@@ -97,6 +102,9 @@ func (s *VolumeService) AttachVolume(ctx context.Context, volumeID string, req V
 
 // DetachVolume detaches a volume - instance must be shut down first
 func (s *VolumeService) DetachVolume(ctx context.Context, volumeID string, req VolumeDetachRequest) error {
+	if err := req.Validate(); err != nil {
+		return err
+	}
 	payload := map[string]interface{}{
 		"id":          volumeID,
 		"action":      VolumeActionDetach,
@@ -108,6 +116,9 @@ func (s *VolumeService) DetachVolume(ctx context.Context, volumeID string, req V
 
 // CloneVolume clones a volume and returns the new volume ID
 func (s *VolumeService) CloneVolume(ctx context.Context, volumeID string, req VolumeCloneRequest) (string, error) {
+	if err := req.Validate(); err != nil {
+		return "", err
+	}
 	actionReq := VolumeActionRequest{
 		ID:     volumeID,
 		Action: VolumeActionClone,
@@ -149,6 +160,9 @@ func (s *VolumeService) CloneVolume(ctx context.Context, volumeID string, req Vo
 
 // ResizeVolume grows a volume - shrinking is not supported
 func (s *VolumeService) ResizeVolume(ctx context.Context, volumeID string, req VolumeResizeRequest) error {
+	if err := req.Validate(); err != nil {
+		return err
+	}
 	actionReq := VolumeActionRequest{
 		ID:     volumeID,
 		Action: VolumeActionResize,
@@ -159,6 +173,9 @@ func (s *VolumeService) ResizeVolume(ctx context.Context, volumeID string, req V
 }
 
 func (s *VolumeService) RenameVolume(ctx context.Context, volumeID string, req VolumeRenameRequest) error {
+	if err := req.Validate(); err != nil {
+		return err
+	}
 	actionReq := VolumeActionRequest{
 		ID:     volumeID,
 		Action: VolumeActionRename,
